requirements.txt
streamlit>=1.46.1
python-dotenv>=1.1.0
sentence-transformers>=4.1.0
numpy>=1.26.4
unstructured[pdf]>=0.10.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-chroma>=0.1.0
langchain-unstructured>=0.1.0
langchain-huggingface>=0.1.0
transformers>=4.35.0
Pillow>=9.0.0
click>=8.0.0
ibm-watsonx-ai>=1.1.17
langchain-ibm>=0.2.1 
